@@ -14,7 +14,7 @@ import {CustomerPickerModule} from "@app/Components";
 import {MatProgressSpinnerModule} from "@angular/material/progress-spinner";
 import {registerLocaleData} from "@angular/common";
 import localeFr from '@angular/common/locales/fr'
-import {HTTP_INTERCEPTORS} from "@angular/common/http";
+import { HTTP_INTERCEPTORS } from "@angular/common/http";
 import {ErrorCatchingInterceptor} from "@app/http-error.interceptor";
 
 
