@@ -1,5 +1,5 @@
 import {Injectable} from "@angular/core";
-import {HttpClient} from "@angular/common/http";
+import { HttpClient } from "@angular/common/http";
 import {Customer} from "../../entities";
 import {SERVER_URL} from "./http-config";
 import {firstValueFrom} from "rxjs";
