@@ -3,7 +3,7 @@ import {UserService} from "./user.service";
 import {AuthenticationService} from "./authentication.service";
 import {ProfileService} from "./profile.service";
 import {AuthorizationInterceptor} from "@app/identity/interceptors";
-import {HTTP_INTERCEPTORS} from "@angular/common/http";
+import { HTTP_INTERCEPTORS } from "@angular/common/http";
 
 @NgModule({
   providers: [ UserService, AuthenticationService, ProfileService,
